requirements.txt
python-dotenv==0.14.0
selenium==3.141.0
urllib3==1.25.10
emoji~=0.6.0